\name{Kalimdor-package}
\alias{Kalimdor-package}
\alias{Kalimdor}
\docType{package}
\title{
\packageTitle{Kalimdor}
}
\description{
\packageDescription{Kalimdor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Kalimdor}
\packageIndices{Kalimdor}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Kalimdor}

Maintainer: \packageMaintainer{Kalimdor}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
